do not propagate "use of closed conn" error if expected
There's unfortunately no way (AFAIK) of actively interrupting a
net.Listener.Accept() call, so we have to deal with it erroring out when
its net.Listener is closed.
Unfortunately(2) there's also no better way of matching the low-level
error than string checks, since it is not exported by the stdlib.